+++ python/CHANGELOG.md
@@ -41,6 +41,7 @@ _At the moment, the project does **not** adhere to [Semantic Versioning](https:/
 - `btc.sign_tx` will not preload transaction data from `prev_txes`, as usage with TxApi
   is being removed
 - PIN protection and passphrase protection for `ping()` command was removed
+- compatibility no-op code from trezorlib 0.9 was removed from `trezorlib.client`
 
 ## [0.11.6] - 2019-12-30
 [0.11.6]: https://github.com/trezor/trezor-firmware/compare/python/v0.11.5...python/v0.11.6